optimise hrf near workingn
