+++ glmdenoise/utils/getcanonicalhrf.py
@@ -1,5 +1,4 @@
 import numpy as np
-import matplotlib.pyplot as plt
 from scipy.interpolate import pchip
